requirements.txt
spacy
https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.2.0/de_core_news_sm-2.2.0.tar.gz #egg=de_core_news_sm==2.2.0
streamlit
sklearn
#sklearn.feature_extraction.text
xgboost
wordcloud
joblib
matplotlib.backends.backend_agg
#matplotlib
#matplotlib.pyplot
re
string
unicodedata
nltk
nltk.corpus
bs4
PIL

